refactor: 키워드 선택 페이지를 컴포넌트 단위로 분리 적용
